requirements.txt
asgiref==3.2.7
Django==3.0.5
django-bootstrap-modal-forms==1.5.0
django-bootstrap3==12.0.3
django-braces==1.14.0
django-mathfilters==1.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2019.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
